Add the ability to theme pages
This commit adds the ability for a specific page or section of the
website to have different themes.

A class is automatically to every page scoped to its category. Add a
theme for a category in `themes.scss`, overriding as few or as many
colors as necessary.

e.g. `.themes--help` would theme all help articles.

You can also theme individual pages by providing a `theme` key and value
in the article [frontmatter].

e.g. `theme: dark` in the article frontmatter would apply a
`themes--dark` class to the page.

[frontmatter]: https://middlemanapp.com/basics/frontmatter/