Markup, styles, logic for plays
This commit adds markup, logic and styles for pages that showcase a
stageplay. Most content sections are opt-in, which allows unproduced and
produced plays to be added. Image galleries, video and audio content
will be addressed in future commits.

This commit introduces a number of CSS components:
- `inline-lists` create dot-separated lists in a continuous line
- `button`s are an opt-in component that styles an element as a button
- `rack`s and `stack`s lay out content horizontally and vertically

Trello: https://trello.com/c/rqFzapKm