View content from external sources
This commit adds the ability to view content archived from external
sources. This is necessary because some of the content on the site
links to Jamaican websites, which do not always migrate older content
at stable URLs. Archiving a copy of the relevant external content allows
the site to maintain the reliability of linking to primary sources, even
if those sources no longer exist.