Landing page layout, refactor section component
This commit refactors the `section` component to be a simple container,
agnostic to its contents. A new `editorial` component is tasked with
laying out editorial content.

It also introduces a new landing page layout, which (supplied with the
appropriate frontmatter) can be used to give a project a
product-oriented look and feel.

Landing page layout