Allow granular sharing of pages
Somewhat counter-intuitively, this commit removes navigation links from
pages and the footer.

The site is mostly an archive of past work. Sometimes it is beneficial
to share a particular design project, or past article, or play with
someone. But it is not beneficial for that person to access the rest of
the site, eg for an op-ed editor to know of my playwriting, or for a
design director to know I construct crosswords.

Closes #80.